requirements.txt
virtualenv
meetup-api
pandas
numpy >= 1.7.0
scikit-learn
scipy
matplotlib
PySide
